Separate Exareme integration tests from other algorithm tests
