+++ Exareme-Docker/src/mip-algorithms/tests/algorithm_tests/test_pca.py
@@ -1,16 +1,10 @@
-import json
-
 import numpy as np
 import pytest
-import requests
 from PCA import PCA
 from mipframework.testutils import get_test_params, get_algorithm_result
 from pathlib import Path
-from tests import vm_url
 
 expected_file = Path(__file__).parent / "expected" / "pca_expected.json"
-headers = {"Content-type": "application/json", "Accept": "text/plain"}
-url = vm_url + "PCA"
 
 
 @pytest.mark.parametrize(
@@ -30,20 +24,6 @@ def test_pca_algorithm_local(test_input, expected):
 )
 def test_pca_algorithm_federated(test_input, expected):
     result = get_algorithm_result(PCA, test_input, num_workers=10)
-
-    assert int(result["n_obs"]) == int(expected["n_obs"])
-    assert np.isclose(result["eigenvalues"], expected["eigen_vals"], atol=1e-3).all()
-    for u, v in zip(result["eigenvectors"], expected["eigen_vecs"]):
-        assert are_collinear(u, v)
-
-
-@pytest.mark.parametrize(
-    "test_input, expected", get_test_params(expected_file, slice(95, 100))
-)
-def test_pca_algorithm_exareme(test_input, expected):
-    result = requests.post(url, data=json.dumps(test_input), headers=headers)
-    result = json.loads(result.text)
-    result = result["result"][0]["data"]
 
     assert int(result["n_obs"]) == int(expected["n_obs"])
     assert np.isclose(result["eigenvalues"], expected["eigen_vals"], atol=1e-3).all()
